src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==5.0.2               # via kombu
appdirs==1.4.4            # via black
asgiref==3.3.1            # via django
attrs==20.3.0             # via flake8-bugbear, flake8-eradicate, pytest
awesome-slugify==1.6.5    # via -r requirements.in
backcall==0.2.0           # via ipython
billiard==3.6.3.0         # via celery
black==20.8b1             # via flake8-black
celery==5.0.4             # via -r requirements.in
certifi==2020.12.5        # via requests, sentry-sdk
chardet==3.0.4            # via requests
click-didyoumean==0.0.3   # via celery
click-plugins==1.1.1      # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via black, celery, click-didyoumean, click-plugins, click-repl, pip-tools
ddtrace==0.44.0           # via -r requirements.in
debtcollector==2.2.0      # via oslo.config, oslo.utils, python-keystoneclient
decorator==4.4.2          # via ipython
django-behaviors==0.5.1   # via -r requirements.in
django-cors-headers==3.5.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-healthchecks==1.4.2  # via -r requirements.in
django-js-asset==1.2.2    # via django-mptt
django-mptt==0.11.0       # via -r requirements.in
django-redis==5.0.0       # via -r requirements.in
django-storage-swift==1.2.19  # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-cors-headers, django-filter, django-healthchecks, django-mptt, django-redis, djangorestframework, djangorestframework-simplejwt, drf-nested-routers
djangorestframework-simplejwt==4.6.0  # via -r requirements.in
djangorestframework==3.12.2  # via -r requirements.in, djangorestframework-simplejwt, drf-nested-routers
drf-nested-routers==0.92.5  # via -r requirements.in
environ==1.0              # via -r requirements.in
eradicate==2.0.0          # via flake8-eradicate
faker==0.9.1              # via mixer
flake8-black==0.2.1       # via -r requirements.in
flake8-bugbear==20.11.1   # via -r requirements.in
flake8-eradicate==1.0.0   # via -r requirements.in
flake8-isort==4.0.0       # via -r requirements.in
flake8-print==4.0.0       # via -r requirements.in
flake8==3.8.4             # via -r requirements.in, flake8-black, flake8-bugbear, flake8-eradicate, flake8-isort, flake8-print
freezegun==1.0.0          # via -r requirements.in, pytest-freezegun
idna==2.8                 # via requests
iniconfig==1.1.1          # via pytest
intervaltree==3.1.0       # via ddtrace
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements.in
iso8601==0.1.13           # via keystoneauth1, oslo.utils
isort==5.6.4              # via -r requirements.in, flake8-isort
jedi==0.17.2              # via ipython
keystoneauth1==4.3.0      # via python-keystoneclient
kombu==5.0.2              # via celery
mccabe==0.6.1             # via flake8
mixer==6.1.3              # via -r requirements.in
msgpack==1.0.0            # via oslo.serialization
mypy-extensions==0.4.3    # via black
netaddr==0.8.0            # via oslo.config, oslo.utils
netifaces==0.10.9         # via oslo.utils
os-service-types==1.7.0   # via keystoneauth1
oslo.config==8.4.0        # via python-keystoneclient
oslo.i18n==5.0.1          # via oslo.config, oslo.utils, python-keystoneclient
oslo.serialization==4.0.1  # via python-keystoneclient
oslo.utils==4.7.0         # via oslo.serialization, python-keystoneclient
packaging==20.7           # via oslo.utils, pytest
parso==0.7.1              # via jedi
pathspec==0.8.1           # via black
pbr==5.5.1                # via debtcollector, keystoneauth1, oslo.i18n, oslo.serialization, oslo.utils, python-keystoneclient, stevedore
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.4.0          # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.8     # via click-repl, ipython
protobuf==3.14.0          # via ddtrace
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8, flake8-print
pyflakes==2.2.0           # via flake8
pygments==2.7.3           # via ipython
pyjwt==1.7.1              # via djangorestframework-simplejwt
pyparsing==2.4.7          # via oslo.utils, packaging
pytest-django==4.1.0      # via -r requirements.in
pytest-env==0.6.2         # via -r requirements.in
pytest-freezegun==0.4.2   # via -r requirements.in
pytest-mock==3.3.1        # via -r requirements.in
pytest==6.1.2             # via -r requirements.in, pytest-django, pytest-env, pytest-freezegun, pytest-mock
python-dateutil==2.8.1    # via faker, freezegun
python-keystoneclient==4.2.0  # via django-storage-swift
python-magic==0.4.18      # via django-storage-swift
python-swiftclient==3.11.0  # via django-storage-swift
pytz==2020.4              # via celery, django, oslo.serialization, oslo.utils
pyyaml==5.3.1             # via oslo.config
redis==3.5.3              # via -r requirements.in, django-redis
regex==2020.11.13         # via awesome-slugify, black
requests-mock==1.8.0      # via -r requirements.in
requests==2.22.0          # via -r requirements.in, django-healthchecks, keystoneauth1, oslo.config, python-keystoneclient, python-swiftclient, requests-mock
rfc3986==1.4.0            # via oslo.config
sentry-sdk==0.19.5        # via -r requirements.in
six==1.15.0               # via click-repl, debtcollector, django-healthchecks, django-storage-swift, faker, flake8-print, keystoneauth1, oslo.i18n, oslo.utils, pip-tools, protobuf, python-dateutil, python-keystoneclient, python-swiftclient, requests-mock, tenacity
sortedcontainers==2.3.0   # via intervaltree
sqlparse==0.4.1           # via django
statsd==3.3.0             # via -r requirements.in
stevedore==3.3.0          # via keystoneauth1, oslo.config, python-keystoneclient
tenacity==6.2.0           # via ddtrace
testfixtures==6.16.0      # via flake8-isort
text-unidecode==1.2       # via faker
toml==0.10.2              # via black, pytest
traitlets==5.0.5          # via ipython
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
unidecode==0.4.21         # via awesome-slugify
urllib3==1.25.11          # via requests, sentry-sdk
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
wrapt==1.12.1             # via debtcollector

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
